% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_metrics.R
\name{compare_metrics}
\alias{compare_metrics}
\title{Compare_Metrics.R}
\usage{
compare_metrics(sig_dir, genes_dir, logT = FALSE, make_chart = FALSE)
}
\arguments{
\item{sig_dir}{This is the signature directory file location}

\item{genes_dir}{This is the gene expression file location}
}
\description{
Compares mean, median, first PCA of gene scores for co-correlation
}
\examples{
compare_metrics()
}
\keyword{expression}
