% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_metrics_loc.R
\name{compare_metrics_loc}
\alias{compare_metrics_loc}
\title{compare_metrics_loc.R}
\usage{
compare_metrics_loc(gene_sigs_list, names_sigs, mRNA_expr_matrix,
  names_datasets, out_dir = "~", file = NULL, showResults = FALSE,
  radar_plot_values)
}
\arguments{
\item{gene_sigs_list}{A list of genes representing the gene signature to be tested.}

\item{names_sigs}{The names of the gene signatures (one name per gene signature, in gene_sigs_list)}

\item{mRNA_expr_matrix}{A list of expression matrices}

\item{names_datasets}{The names of the different datasets contained in mRNA_expr_matrix}

\item{out_dir}{A path to the directory where the resulting output files are written}

\item{file}{File representing the log file where errors can be written}

\item{showResults}{Tells if open dialog boxes showing the computed results. Default is FALSE}

\item{radar_plot_values}{A list of values that store computations that will be used in the final summary radarplot}
}
\description{
This function creates the plots to compare various signature summary statistic metrics against each other 
That is, compares the mean, median and first principal component with each other, and also produces PCA vs variance
plot for the first 10 principal components of the dataset
}
\keyword{compare_metrics_loc}
