% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.heatmaps.R
\name{draw.heatmaps}
\alias{draw.heatmaps}
\title{draw.heatmaps.R}
\usage{
draw.heatmaps(hmaps, names_datasets, names_sigs)
}
\arguments{
\item{hmaps}{A list of genes representing the gene signature to be tested.}

\item{names_datasets}{A list of names of the datasets (for sizing purposes only)}

\item{names_sigs}{A list of names of gene signatures input (for sizing purposes only)}
}
\description{
This function draws the list of heatmaps given to it in a grid on the R graphics area
grid size has number of rows as the number of gene signatures given to the code, and 
number of columns is the number of datasets on which to test the signature
}
\keyword{draw.heatmaps}
