% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_compactness.R
\name{eval_compactness}
\alias{eval_compactness}
\title{Eval_Compactness.R}
\usage{
eval_compactness(sig_dir, genes_dir)
}
\arguments{
\item{sig_dir}{This is the signature directory file location}

\item{genes_dir}{This is the gene expression file location}
}
\description{
Produces a heatmap of an autocorrelation matrix of genes involved in a gene signature, which can be used to evaluate signature compactness.
}
\examples{
eval_compactness()
}
\keyword{compactness}
