% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_compactness_loc.R
\name{eval_compactness_loc}
\alias{eval_compactness_loc}
\title{eval_compactness_loc.R}
\usage{
eval_compactness_loc(gene_sigs_list, names_sigs, mRNA_expr_matrix,
  names_datasets, out_dir = "~", file = NULL, showResults = FALSE,
  radar_plot_values)
}
\arguments{
\item{gene_sigs_list}{A list of genes representing the gene signature to be tested.}

\item{names_sigs}{The names of the gene signatures (one name per gene signature, in gene_sigs_list)}

\item{mRNA_expr_matrix}{A list of expression matrices}

\item{names_datasets}{The names of the different datasets contained in mRNA_expr_matrix}

\item{out_dir}{A path to the directory where the resulting output files are written}

\item{file}{File representing the log file where errors can be written}

\item{showResults}{Tells if open dialog boxes showing the computed results. Default is FALSE}

\item{radar_plot_values}{A list of values that store computations that will be used in the final summary radarplot}
}
\description{
This function creates the plots of autocorrelation, as well as the rank product computation. Specifically, it creates
plots for heatmap of autocorrelation, density plot of autocorrelation values, and if there is more than one dataset, for
each individual gene signature, will do a rank product analysis on the lists of median autocorrelation coefficients to 
identify genes consistently low ranking in autocorrelation among more than one dataset.
}
\keyword{eval_compactness_loc}
