% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_expr.R
\name{eval_expr}
\alias{eval_expr}
\title{Eval_Expr.R}
\usage{
eval_expr(sig_dir, genes_dir, make_chart = FALSE)
}
\arguments{
\item{sig_dir}{This is the signature directory file location}

\item{genes_dir}{This is the gene expression file location}
}
\description{
Evaluates the expression of the signature genes across samples
}
\examples{
eval_expr()
}
\keyword{expression}
