% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_expr_loc.R
\name{eval_expr_loc}
\alias{eval_expr_loc}
\title{eval_expr_loc.R}
\usage{
eval_expr_loc(gene_sigs_list, names_sigs, mRNA_expr_matrix, names_datasets,
  thresholds = NULL, out_dir = "~", file = NULL, showResults = FALSE,
  radar_plot_values)
}
\arguments{
\item{gene_sigs_list}{A list of genes representing the gene signature to be tested.}

\item{names_sigs}{The names of the gene signatures (one name per gene signature, in gene_sigs_list)}

\item{mRNA_expr_matrix}{A list of expression matrices}

\item{names_datasets}{The names of the different datasets contained in mRNA_expr_matrix}

\item{thresholds}{A list of thresholds to be considered for each data set, default is median of the data set. A gene is considered expressed if above the threshold, non-expressed otherwise. One threshold per dataset, in the same order as the dataset list.}

\item{out_dir}{A path to the directory where the resulting output files are written}

\item{file}{File representing the log file where errors can be written}

\item{showResults}{Tells if open dialog boxes showing the computed results. Default is FALSE}

\item{radar_plot_values}{A list of values that store computations that will be used in the final summary radarplot}
}
\description{
This function creates the plots of measures of expression. Takes in the list of thresholds of expression, and if this is not
specified by the user, uses the median of all genes across all samples as the threshold cutoff for the barchart presenting 
proportion of samples expressing above the threshold. Also produces similar barchart for proportion of samples with NA values 
for each of the genes considered. Creates density plots as well for these barcharts for easy viewing of distribution in the case
of a large number of signature genes.
}
\keyword{eval_expr_loc}
