% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_stan.R
\name{eval_stan}
\alias{eval_stan}
\title{Eval_Stan.R}
\usage{
eval_stan(sig_dir, genes_dir, logT = FALSE, make_chart = FALSE)
}
\arguments{
\item{sig_dir}{This is the signature directory file location}

\item{genes_dir}{This is the gene expression file location}
}
\description{
Evaluates the effect of standardisation of data on the gene signature score.
}
\examples{
eval_stan()
}
\keyword{expression}
