% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_struct.R
\name{eval_struct}
\alias{eval_struct}
\title{Eval_Struct.R}
\usage{
eval_struct(sig_dir, genes_dir, logT = FALSE)
}
\arguments{
\item{sig_dir}{This is the signature directory file location}

\item{genes_dir}{This is the gene expression file location}
}
\description{
Evaluates the structure of the signature when considering expression across patients. Enables visual clustering of signature genes/patients.
}
\examples{
eval_expr()
}
\keyword{expression}
