% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_struct_loc.R
\name{eval_struct_loc}
\alias{eval_struct_loc}
\title{eval_struct_loc.R}
\usage{
eval_struct_loc(gene_sigs_list, names_sigs, mRNA_expr_matrix, names_datasets,
  covariates, out_dir = "~", file = NULL, showResults = FALSE,
  radar_plot_values)
}
\arguments{
\item{gene_sigs_list}{A list of genes representing the gene signature to be tested.}

\item{names_sigs}{The names of the gene signatures (one name per gene signature, in gene_sigs_list)}

\item{mRNA_expr_matrix}{A list of expression matrices}

\item{names_datasets}{The names of the different datasets contained in mRNA_expr_matrix}

\item{covariates}{A list containing a sub-list of 'annotations' and 'colors' which contains the annotation matrix for the given dataset and the associated colours with which to plot in the expression heatmap}

\item{out_dir}{A path to the directory where the resulting output files are written}

\item{file}{File representing the log file where errors can be written}

\item{showResults}{Tells if open dialog boxes showing the computed results. Default is FALSE}

\item{radar_plot_values}{A list of values that store computations that will be used in the final summary radarplot}
}
\description{
This function creates the plots to evaluate for signature structure. Produces plots for the 
expression of the signature with hierarchical clustering as well as biclustering plots on binarized data and on non-binarized data
}
\keyword{eval_struct_loc}
