% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_var.R
\name{eval_var}
\alias{eval_var}
\title{Eval_Var.R}
\usage{
eval_var(sig_dir, genes_dir, make_chart = FALSE)
}
\arguments{
\item{sig_dir}{This is the signature directory file location}

\item{genes_dir}{This is the gene expression file location}
}
\description{
Evaluates the variance of the signature genes across samples
}
\examples{
eval_var()
}
\keyword{expression}
