% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_var_loc.R
\name{eval_var_loc}
\alias{eval_var_loc}
\title{eval_var_loc.R}
\usage{
eval_var_loc(gene_sigs_list, names_sigs, mRNA_expr_matrix, names_datasets,
  out_dir = "~", file = NULL, showResults = FALSE, radar_plot_values)
}
\arguments{
\item{gene_sigs_list}{A list of genes representing the gene signature to be tested.}

\item{names_sigs}{The names of the gene signatures (one name per gene signature, in gene_sigs_list)}

\item{mRNA_expr_matrix}{A list of expression matrices}

\item{names_datasets}{The names of the different datasets contained in mRNA_expr_matrix}

\item{out_dir}{A path to the directory where the resulting output files are written}

\item{file}{File representing the log file where errors can be written}

\item{showResults}{Tells if open dialog boxes showing the computed results. Default is FALSE}

\item{radar_plot_values}{A list of values that store computations that will be used in the final summary radarplot}
}
\description{
This function creates the plots for the measures of variance. It produces the mean vs sd plot, and computes metrics
that are used in the final radar plot such as proportion of signature genes lying in the top 10%, 25%, and 50% of 
overall coefficient of variance across all genes, and the fractional ratio of the skewness of the expression distribution 
of signature genes vs all genes. Also outputs tables containing mean and sd values for each of the signature genes
}
\keyword{eval_var_loc}
