% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_grob.R
\name{grab_grob}
\alias{grab_grob}
\title{grab_grob.R}
\usage{
grab_grob()
}
\description{
This function "grabs" the plots on the R graphics device, and enables them to be saved to an
external file.
}
\keyword{grab_grob}
