% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_radar_chart_loc.R
\name{make_radar_chart_loc}
\alias{make_radar_chart_loc}
\title{make_radar_chart_loc.R}
\usage{
make_radar_chart_loc(radar_plot_values, showResults = FALSE, names_sigs,
  names_datasets, out_dir = "~", file)
}
\arguments{
\item{radar_plot_values}{A list of values that store computations that will be used in the final summary radarplot}

\item{showResults}{Tells if open dialog boxes showing the computed results. Default is FALSE}

\item{names_sigs}{The names of the gene signatures (one name per gene signature, in gene_sigs_list)}

\item{names_datasets}{The names of the different datasets contained in mRNA_expr_matrix}

\item{out_dir}{A path to the directory where the resulting output files are written}

\item{file}{File representing the log file where errors can be written}
}
\description{
This function creates the final summary radar plot, adding in all of the metrics that have been pre-computed
in each of the functions called before it in the code. Also computes the ratio of the area contained within 
each of the signature/dataset radar plots to the ratio of the area of the full polygon, and outputs these proportions
within the legend, as a potential summarizing metric that can be used to compare signatures/datasets in a more quantitative fashion.
}
\keyword{make_radar_chart_loc}
