% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_all_plots.R
\name{make_all_plots}
\alias{make_all_plots}
\title{make_all_plots.R}
\usage{
make_all_plots(sig_dir, genes_dir, logT = FALSE)
}
\arguments{
\item{sig_dir}{This is the signature directory file location}

\item{genes_dir}{This is the gene expression file location}
}
\description{
Makes the plots in each of the subfunctions.
}
\examples{
make_all_plots()
}
\keyword{expression}
